Article details & router functionality
I had to add the router functionality to test that the information was being transfered through the URL.

I added id property in the interface and controllers to transfer the id in order to show the information of each article. Also, I created a button to act as a link, later I can create a card instead of the button.

I had to use image_data and image_media_type to show the image in the article. It keeps showing a message as "unknown data" in the console but it still shows the image. I haven't managed to do so in the email list component.

You may have a lot of changes in the code but it's due to a formatter I use, I believe there won't be any problem.